+++ src/main/java/com/uwaterloo/connect/service/SignUpService.java
@@ -4,7 +4,7 @@ import com.uwaterloo.connect.enums.UserRole;
 import com.uwaterloo.connect.model.EmailToken;
 import com.uwaterloo.connect.model.User;
 import com.uwaterloo.connect.utils.InputValidator;
-import com.uwaterloo.connect.dto.RegistrationRequest;
+import com.uwaterloo.connect.dto.SignUpRequest;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -13,23 +13,23 @@ import java.time.LocalDateTime;
 
 @Service
 @AllArgsConstructor
-public class RegistrationService {
+public class SignUpService {
 
     private final UserService userService;
     private final EmailTokenService emailTokenService;
     private final EmailSender emailSender;
 
 
-    public String register(RegistrationRequest request) {
+    public String signUp(SignUpRequest request) {
         //TODO create custom exceptions
         if (!InputValidator.emailValidator(request.getEmail())) {
             throw new IllegalArgumentException("email not valid");
         }
         String token = userService.signUpUser(new User(request.getFirstName(), request.getLastName(), request.getUserName(), request.getEmail(), request.getPassword(), request.getDateOfBirth(), UserRole.USER));
-        String link = "http://localhost:8080/api/v1/registration/confirm?token=" + token;
+        String link = "http://localhost:8080/api/v1/signup/confirm?token=" + token;
         emailSender.send(request.getEmail(), buildEmail(request.getFirstName(), link));
 
-        return token;
+        return "Success!  Please, check your email for to complete your signup process";
     }
 
     @Transactional
@@ -108,7 +108,7 @@ public class RegistrationService {
                 "      <td width=\"10\" valign=\"middle\"><br></td>\n" +
                 "      <td style=\"font-family:Helvetica,Arial,sans-serif;font-size:19px;line-height:1.315789474;max-width:560px\">\n" +
                 "        \n" +
-                "            <p style=\"Margin:0 0 20px 0;font-size:19px;line-height:25px;color:#0b0c0c\">Hi " + name + ",</p><p style=\"Margin:0 0 20px 0;font-size:19px;line-height:25px;color:#0b0c0c\"> Thank you for registering. Please click on the below link to activate your account: </p><blockquote style=\"Margin:0 0 20px 0;border-left:10px solid #b1b4b6;padding:15px 0 0.1px 15px;font-size:19px;line-height:25px\"><p style=\"Margin:0 0 20px 0;font-size:19px;line-height:25px;color:#0b0c0c\"> <a href=\"" + link + "\">Activate Now</a> </p></blockquote>\n Link will expire in 15 minutes. <p>See you soon</p>" +
+                "            <p style=\"Margin:0 0 20px 0;font-size:19px;line-height:25px;color:#0b0c0c\">Hi " + name + ",</p><p style=\"Margin:0 0 20px 0;font-size:19px;line-height:25px;color:#0b0c0c\"> Thank you for signing up. Please click on the below link to activate your account: </p><blockquote style=\"Margin:0 0 20px 0;border-left:10px solid #b1b4b6;padding:15px 0 0.1px 15px;font-size:19px;line-height:25px\"><p style=\"Margin:0 0 20px 0;font-size:19px;line-height:25px;color:#0b0c0c\"> <a href=\"" + link + "\">Activate Now</a> </p></blockquote>\n Link will expire in 15 minutes. <p>See you soon</p>" +
                 "        \n" +
                 "      </td>\n" +
                 "      <td width=\"10\" valign=\"middle\"><br></td>\n" +
